Bump to 2.0.8, add license headers, and reorganize test packages (#33)
- Bump version from 2.0.7 to 2.0.8 across build.gradle.kts, README.md, CLAUDE.md, and website docs
- Add Apache 2.0 license headers to all source and resource files
- Reorganize test classes from flat com.srcref into subpackages by concern:
  api/ (ApiTest), core/ (FileTests, UrlsTest, ContentCacheTest, QueryParamsTest,
  EndpointsTest, MainTest), pages/ (CommonTest), integration/ (RoutesIntegrationTest),
  regression/ (BugFixTests)
- Remove stale llm-output/claude-bugs-2-14-26.md
